refactor(content): move posts, authors, and series queries out of markdown.ts
content/posts.ts owns post/page discovery, content/authors.ts the author
aggregation and slug derivation, content/series.ts the series queries.
Collections live inside series.ts rather than their own module — a
collection IS a series folder (type: collection): getAllSeries expands
collections via getCollectionPosts while getCollectionPosts resolves its
items via getSeriesPosts/getSeriesData, so a separate collections module
would be an import cycle. All env-keyed Map caches become createMemo/
createKeyedMemo instances. ~40 consumers update imports. New integration
coverage: draft posts excluded in production, future-dated posts excluded
under showFuturePosts:false, and the mixed-format series throw — all
previously untested strict-build behavior.
