+++ scripts/generate-knowledge-graph.ts
@@ -9,7 +9,8 @@
 
 import fs from 'fs';
 import path from 'path';
-import { getAllPosts, getSeriesData } from '../src/lib/markdown';
+import { getSeriesData } from '../src/lib/content/series';
+import { getAllPosts } from '../src/lib/content/posts';
 import { getAllNotes } from '../src/lib/content/notes';
 import { getAllFlows } from '../src/lib/content/flows';
