+++ src/app/archive/page.tsx
@@ -1,5 +1,6 @@
 import Link from 'next/link';
-import { getAllPosts, getSeriesData } from '@/lib/markdown';
+import { getSeriesData } from '@/lib/content/series';
+import { getAllPosts } from '@/lib/content/posts';
 import type { PostData } from '@/lib/content/types';
 import { siteConfig } from '../../../site.config';
 import { resolveLocale, t } from '@/lib/i18n';
