+++ src/app/sitemap.ts
@@ -1,5 +1,5 @@
 import { MetadataRoute } from 'next';
-import { getAllPosts, getAllPages } from '@/lib/markdown';
+import { getAllPosts, getAllPages } from '@/lib/content/posts';
 import { getAllFlows } from '@/lib/content/flows';
 import { getAllBooks } from '@/lib/content/books';
 import { siteConfig } from '../../site.config';
